requirements.txt
pip==21.3.1
wheel==0.37.1
requests==2.28.2
colorama==0.4.6
attrs==22.2.0
mysql-connector-python==8.0.32
pytest==7.2.1
selenium==3.141.0
Appium-Python-Client==1.3.0
urllib3==1.26.14
protobuf==3.20.3
pluggy==1.0.0
exceptiongroup==1.1.0
iniconfig==2.0.0
certifi==2022.12.7
six==1.16.0
idna==3.4
websocket-client==0.59.0
eyes-selenium==5.16.0
marshmallow==2.21.0
setuptools==60.2.0
tqdm==4.64.1
python-dotenv==0.21.1
packaging==23.0